+++ dev/scripts/gmi.js
@@ -135,6 +135,14 @@ var GMI = function(options, embedVars, gameDir) {
         return false;
     };
     GMI.prototype.showSettings = function() {
+        var settingsDiv = document.getElementsByClassName("settings");
+        if (!(settingsDiv && settingsDiv[0])) {
+            var settings = document.createElement('div');
+            settings.className = "settings"
+            settings.innerHTML += "The settings screen will appear here when the game is hosted on the BBC servers (click here to close)";
+            settings.addEventListener("click", function() { document.body.removeChild(settings); });
+            document.body.appendChild(settings);
+        }
         return true;
     };
     GMI.prototype.sendStatsEvent = function(name, type, params) {