+++ test/core/startup.spec.js
@@ -7,6 +7,7 @@ import { createMockGmi } from "../fake/gmi";
 import { domElement } from "../fake/dom-element";
 
 import { startup } from "../../src/core/startup.js";
+import { getBrowser } from "../../src/core/browser.js";
 import * as gmiModule from "../../src/core/gmi/gmi.js";
 import * as Game from "../fake/game.js";
 import * as Scene from "../../src/core/scene.js";
@@ -14,6 +15,10 @@ import * as LoadFonts from "../../src/core/font-loader.js";
 import * as Navigation from "../../src/core/navigation.js";
 import * as styles from "../../src/core/custom-styles.js";
 import * as qaMode from "../../src/core/qa/qa-mode.js";
+import * as a11y from "../../src/core/accessibility/accessibility-layer.js";
+import * as fullscreen from "../../src/core/fullscreen.js";
+
+jest.mock("../../src/core/browser.js");
 
 describe("Startup", () => {
     let mockGmi;
@@ -35,6 +40,7 @@ describe("Startup", () => {
                 return containerDiv;
             }
         });
+        getBrowser.mockImplementation(() => ({ forceCanvas: false, isSilk: false }));
         jest.spyOn(styles, "addCustomStyles");
         jest.spyOn(Phaser, "Game").mockImplementation(() => Game.Stub);
         global.window.getGMI = jest.fn().mockImplementation(() => mockGmi);
@@ -59,11 +65,12 @@ describe("Startup", () => {
         const expectedConfig = {
             width: 1400,
             height: 600,
-            renderer: 1,
+            renderer: 0,
             antialias: true,
-            multiTexture: false,
+            multiTexture: true,
             parent: containerDiv,
-            transparent: true,
+            transparent: false,
+            clearBeforeRender: false,
         };
 
         const actualConfig = Phaser.Game.mock.calls[0][0];
@@ -75,6 +82,25 @@ describe("Startup", () => {
         expect(actualConfig.multiTexture).toBe(expectedConfig.multiTexture);
         expect(actualConfig.parent).toEqual(expectedConfig.parent);
         expect(actualConfig.transparent).toBe(expectedConfig.transparent);
+        expect(actualConfig.transparent).toBe(expectedConfig.transparent);
+        expect(actualConfig.clearBeforeRender).toBe(expectedConfig.clearBeforeRender);
+    });
+
+    it("sets transparent config flag to false when Amazon Silk Browser", () => {
+        const mockSilkBrowser = { name: "Amazon Silk", isSilk: true, version: "1.1.1" };
+        getBrowser.mockImplementation(() => mockSilkBrowser);
+
+        startup();
+        const actualConfig = Phaser.Game.mock.calls[0][0];
+        expect(actualConfig.transparent).toBe(true);
+    });
+
+    it("sets renderer to canvas when browser returns forceCanvas", () => {
+        const mockSafari9 = { name: "Safari", forceCanvas: true };
+        getBrowser.mockImplementation(() => mockSafari9);
+        startup();
+        const actualConfig = Phaser.Game.mock.calls[0][0];
+        expect(actualConfig.renderer).toBe(1);
     });
 
     test("throws an error if the game container element cannot be found", () => {
@@ -88,6 +114,8 @@ describe("Startup", () => {
             jest.spyOn(LoadFonts, "loadFonts").mockImplementation(() => {});
             jest.spyOn(Navigation, "create").mockImplementation(() => {});
             jest.spyOn(qaMode, "create").mockImplementation(() => {});
+            jest.spyOn(a11y, "setup").mockImplementation(() => {});
+            jest.spyOn(fullscreen, "listenForTap").mockImplementation(() => {});
 
             startup({}, "NavConfig");
             const game = Phaser.Game.mock.calls[0][0];
@@ -131,6 +159,18 @@ describe("Startup", () => {
             const statsParams = gmiModule.startStatsTracking.mock.calls[0];
             expect(statsParams[0]).toEqual(Game.Stub);
             expect(JSON.stringify(statsParams[1])).toEqual(JSON.stringify(expectedContext));
+        });
+
+        it("sets up the accessibility manager", () => {
+            const onComplete = LoadFonts.loadFonts.mock.calls[0][1];
+            onComplete();
+            expect(a11y.setup).toHaveBeenCalledWith(Game.Stub.canvas.parentElement);
+        });
+
+        it("sets up the fullscreen API", () => {
+            const onComplete = LoadFonts.loadFonts.mock.calls[0][1];
+            onComplete();
+            expect(fullscreen.listenForTap.mock.calls[0]).toEqual([Game.Stub.canvas.parentElement, Game.Stub]);
         });
     });
 });