+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 | Version | Description |
 |---------|-------------|
 | | Migrates unit tests from Mocha/Chai/Sinon to Jest. |
+| | Configure webpack-dev-server to store server output on disk. |
+| | Changes fullscreen target for Phaser to ensure the settings screen and Brim appear above the fullscreen game, adds mock GMI for local testing. |
+| 1.0.7 | |
+| | Add lib folder to babel load in webpack config. |
+| | Lock Prettier to version in package.json. |
+| | Npm Audit fixes. |
+| | Add checks to force canvas rendering on ipad 2. |
+| | Enable multi-texture support. |
+| | Set clearBeforeRender to false since all games fill the canvas. |
+| | Set transparent canvas (Kindle flicker fix) only if Amazon Silk browser. |
 | | QAMode now occurs when on a test URL. |
 | 1.0.6 | |
 | | Add removeFromAccessibleButtons and getAccessibleButtons to accessible layer. |